updates
changed the scale of the flowers so each one of them are not the same, changed  the source size of the text and made it appear bigger by drawing the same pixel  twice bigger than the original.(took me two days to complete and explain to myself what the guy who made the template tried to do with it ).
tried to calculate how fast is the bee going in diagonal so i can make it slower but it didn't quite  work so maybe i can ask someone later.
Also added comments for every line of code so not only me can understand what i tried to implement with them!!!